Add: latest release download (#8)
* Add: upload-and-release.yml (#2)

Description:
Handles automatic deployment of releases for us. Triggers on push/merge to the Release branch

* Update build-and-upload.yml

* Fix: refactored logic for backgroundworker process | Fix: UI Design
Description: Added logic and related drop-down list for user to select a specific process/window they want to send keys to. This allows users to specify that only the game will get keypresses and nothing else. Also ensures our compatability with most games that block automated input/key-sends.

* Fix: refactored logic for backgroundworker process | Fix: UI Design (#4)

Description: Added logic and related drop-down list for user to select a specific process/window they want to send keys to. This allows users to specify that only the game will get keypresses and nothing else. Also ensures our compatability with most games that block automated input/key-sends.

* Fix: Issue #5 | Delete: Unused classes and resources
Description: Issue with lingering process has been resolved. All resources are now released when the app is closed, but it does sometimes take a bit of time for the backgroundworker to finish. In future, we need to add a message to tell user to wait a bit while the worker finishes

* Fix: Issue #7 | Fix: Logic for play btn's backgroundworker
Description: Resolved issue with song pause/replay and related progressbar issue. Resolved issue where error was thrown from spawning a duplicate backgroundworker. We now only spawn a new worker when the old one is finished.

* Delete: Legacy version of the app
Description: Only the advanced/upgraded version of the app is now available. This improves user experience and app usability